requirements.txt
Flask==3.1.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opencv-contrib-python==4.10.0.84
ultralytics==8.3.9
torchvision==0.19.0
torch==2.4.0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
matplotlib==3.9.2
pillow==10.4.0
numpy==1.26.4
inference-sdk==0.36.1
pytesseract==0.3.13
easyocr==1.7.2
Levenshtein==0.25.1

